# blueprint/skeleton-example/contract-example/src/test/resources/openapi-baseline.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "contract-example",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/v1/orders" : {
      "post" : {
        "operationId" : "createOrder",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "order-controller" ]
      }
    },
    "/v1/orders/{id}" : {
      "get" : {
        "operationId" : "getOrder",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "order-controller" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "CreateOrderRequest" : {
        "type" : "object",
        "properties" : {
          "customerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "note" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "quantity" : {
            "maximum" : 1000,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32"
          },
          "sku" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "customerId", "quantity", "sku" ]
      },
      "OrderResponse" : {
        "type" : "object",
        "properties" : {
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "customerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "note" : {
            "type" : "string"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sku" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "PENDING", "CONFIRMED", "CANCELLED" ]
          }
        },
        "required" : [ "createdAt", "customerId", "id", "quantity", "sku", "status" ]
      }
    }
  }
}